+++ GhettoStation/UAVTalk.cpp
@@ -388,8 +388,8 @@ int uavtalk_read(void) {
 				case GPSPOSITIONSENSOR_OBJID:
 					uav_lat			= uavtalk_get_int32(&msg, GPSPOSITION_OBJ_LAT) / 10000000.0;
 					uav_lon			= uavtalk_get_int32(&msg, GPSPOSITION_OBJ_LON) / 10000000.0;
-					uav_satellites_visible	= uavtalk_get_int8(&msg, GPSPOSITION_OBJ_SATELLITES);
-					uav_fix_type		= uavtalk_get_int8(&msg, GPSPOSITION_OBJ_STATUS);
+					uav_satellites_visible	= (uint8_t) uavtalk_get_int8(&msg, GPSPOSITION_OBJ_SATELLITES);
+					uav_fix_type		= (uint8_t) uavtalk_get_int8(&msg, GPSPOSITION_OBJ_STATUS);
 					uav_gpsheading		= (int16_t) uavtalk_get_float(&msg, GPSPOSITION_OBJ_HEADING);
                                         #ifndef BARO_ALT
 					uav_alt			= uavtalk_get_float(&msg, GPSPOSITION_OBJ_ALTITUDE);
